src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-disable-expected.txt

Tests to ensure that shadow element cannot be created in elements having dynamically created shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

DIV
FAIL element.createShadowRoot() should not be null. Threw exception TypeError: element.createShadowRoot is not a function
SPAN
FAIL element.createShadowRoot() should not be null. Threw exception TypeError: element.createShadowRoot is not a function
A
FAIL element.createShadowRoot() should not be null. Threw exception TypeError: element.createShadowRoot is not a function
SECTION
FAIL element.createShadowRoot() should not be null. Threw exception TypeError: element.createShadowRoot is not a function
g
FAIL element.createShadowRoot() should not be null. Threw exception TypeError: element.createShadowRoot is not a function
rect
FAIL element.createShadowRoot() should not be null. Threw exception TypeError: element.createShadowRoot is not a function
svg
FAIL element.createShadowRoot() should not be null. Threw exception TypeError: element.createShadowRoot is not a function
INPUT
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
EMBED
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
OBJECT
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
AUDIO
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
VIDEO
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
SELECT
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
IMG
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
TEXTAREA
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
IFRAME
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
CANVAS
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
METER
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
PROGRESS
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
VIDEO
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
AUDIO
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
FIELDSET
PASS element.createShadowRoot() threw exception TypeError: element.createShadowRoot is not a function.
PASS successfullyParsed is true

TEST COMPLETE
